+++ redisinsight/ui/src/pages/browser/modules/key-details/components/array-details/array-tabs/ArrayTabs.spec.tsx
@@ -2,11 +2,20 @@ import React from 'react'
 import { fireEvent, render, screen } from 'uiSrc/utils/test-utils'
 
 import ArrayTabs from './ArrayTabs'
+import { ArrayTabsProps } from './ArrayTabs.types'
 import { ARRAY_DETAILS_TAB_LABELS, ArrayDetailsTab } from '../constants'
 
 describe('ArrayTabs', () => {
+  const defaultProps: ArrayTabsProps = {
+    value: ArrayDetailsTab.View,
+    onChange: jest.fn(),
+  }
+
+  const renderComponent = (propsOverride?: Partial<ArrayTabsProps>) =>
+    render(<ArrayTabs {...defaultProps} {...propsOverride} />)
+
   it('renders View, Aggregate and Search tabs with the active one selected', () => {
-    render(<ArrayTabs value={ArrayDetailsTab.View} onChange={jest.fn()} />)
+    renderComponent()
 
     expect(screen.getByTestId('array-tabs')).toBeInTheDocument()
 
@@ -19,7 +28,7 @@ describe('ArrayTabs', () => {
 
   it('calls onChange with the picked tab id', () => {
     const onChange = jest.fn()
-    render(<ArrayTabs value={ArrayDetailsTab.View} onChange={onChange} />)
+    renderComponent({ onChange })
 
     fireEvent.mouseDown(screen.getByText(ARRAY_DETAILS_TAB_LABELS.aggregate))
