+++ redisinsight/ui/src/pages/browser/modules/key-details/components/array-details/array-tabs/ArrayTabs.types.ts
@@ -1,0 +1,6 @@
+import { ArrayDetailsTab } from '../constants'
+
+export interface ArrayTabsProps {
+  value: ArrayDetailsTab
+  onChange: (tab: ArrayDetailsTab) => void
+}
