+++ redisinsight/ui/src/pages/browser/modules/key-details/components/array-details/array-tabs/index.ts
@@ -1,4 +1,4 @@
 import ArrayTabs from './ArrayTabs'
 
 export default ArrayTabs
-export type { ArrayTabsProps } from './ArrayTabs'
+export type { ArrayTabsProps } from './ArrayTabs.types'
