+++ redisinsight/ui/src/pages/browser/modules/key-details/components/array-details/view-tab/ViewTab.types.ts
@@ -1,0 +1,5 @@
+import { RedisResponseBuffer } from 'uiSrc/slices/interfaces'
+
+export interface ViewTabProps {
+  keyProp: RedisResponseBuffer | null
+}
