+++ redisinsight/ui/src/pages/browser/modules/key-details/components/array-details/view-tab/index.ts
@@ -1,4 +1,4 @@
 import ViewTab from './ViewTab'
 
 export default ViewTab
-export type { ViewTabProps } from './ViewTab'
+export type { ViewTabProps } from './ViewTab.types'
